Add support for alternative crypto library (Libgcrypt + sha1-git) (#1012)
This patch adds a wrapper that translates the OpenSSL functions used on
aircrack-ng to the Libgcrypt counterparts.

For the SHA1 functions instead of relying on Libgcrypt it uses an internal
SHA1 Implementation that was borrowed from GIT. This is done because the
SHA1 Libgcrypt implementation is slow and the GIT SHA1 Implementation is
faster than any other.

If you compare the speed cracking a wpa key with this patch (make gcrypt=true)
and without it, you will find that it performs almost at the same speed.

On SSE2 enabled CPUs aircrack-ng will perform like a ~4% faster with OpenSSL,
and on the rest of CPUs (including non-x86 archs) aircrack-ng will perform
like a ~3% faster with this patch.

The explanation for this is:

  * The sha1-git implementation (which is only used on non-SSE2 archs) is
    faster than the OpenSSL one.
  * OpenSSL is faster than Libgcrypt.
  * The sha1-git code is only used on the function (crypto.c:calc_pmk).
    However when running on a SSE2 CPU this code is not loaded, and instead
    a SSE2 optimized version is prefered (sha1-sse2.h:calc_4pmk).

All functions of aircrack-ng were OpenSSL code was used [1] were tested to
perform exactly the same (same input=>same output) when running with OpenSSL
than with this patch (Libgcrypt+SHA1-git)

 [1] crypto.c:{encrypt_wep,calc_pmk,calc_mic,calc_ptk}
     sha1-see2.h:calc_4pmk



git-svn-id: http://svn.aircrack-ng.org/trunk@2169 28c6078b-6c39-48e3-add9-af49d547ecab