osdep: Add basic support for timestamp and link-layer type.
This commit adds very basic support for the packet's timestamp
and data link-layer type. No code currently uses the link-layer,
yet; however, it has been defaulted to a value that matches
current code logic.

As a test, `wpaclean` was extended to support the new fields;
and now correctly preserves the timestamps of cleaned capture
outputs.