libac: Implement retrieving HANDLE from pthread_t value.
This patch performs a hack into the Cygwin internals. A Cygwin
pthread is implemented in C++. The second member variable is
our desired HANDLE value. However, the class has a vtbl as the
first member variable, so we increment over ptrdiff_t. Next member
value is a magic number, as a uint32_t. Next we have our HANDLE.